Moving part around for multiple connection support
