+++ components/awox_mesh/__init__.py
@@ -10,13 +10,16 @@ DEPENDENCIES = ["mqtt", "esp32"]
 awox_ns = cg.esphome_ns.namespace("awox_mesh")
 
 Awox = awox_ns.class_("AwoxMesh", esp32_ble_tracker.ESPBTDeviceListener, cg.Component)
-MeshDevice = awox_ns.class_("MeshDevice", esp32_ble_client.BLEClientBase)
+MeshConnection = awox_ns.class_("MeshConnection", esp32_ble_client.BLEClientBase)
 
 CONNECTION_SCHEMA = esp32_ble_tracker.ESP_BLE_DEVICE_SCHEMA.extend(
     {
-        cv.GenerateID(): cv.declare_id(MeshDevice),
+        cv.GenerateID(): cv.declare_id(MeshConnection),
     }
 ).extend(cv.COMPONENT_SCHEMA)
+
+CONF_CONNECTIONS = "connections"
+MAX_CONNECTIONS = 3
 
 DEVICE_TYPES = {
     "RGB": 0x01,
@@ -25,14 +28,26 @@ DEVICE_TYPES = {
     "PLUG": 0x04,
 }
 
-CONFIG_SCHEMA = (
+def validate_connections(config):
+    if CONF_CONNECTIONS not in config:
+        conf = config.copy()
+        conf[CONF_CONNECTIONS] = [
+            CONNECTION_SCHEMA({}) for _ in range(MAX_CONNECTIONS)
+        ]
+        return conf
+    return config
+
+CONFIG_SCHEMA = cv.All(
     cv.Schema(
         {
             cv.GenerateID(): cv.declare_id(Awox),
             cv.Required("mesh_name"): cv.string_strict,
             cv.Required("mesh_password"): cv.string_strict,
-            cv.Required("address_prefix"): cv.string_strict,
-            cv.Optional("connection", {}): CONNECTION_SCHEMA,
+            cv.Optional("address_prefix"): cv.string_strict,
+            cv.Optional(CONF_CONNECTIONS): cv.All(
+                cv.ensure_list(CONNECTION_SCHEMA),
+                cv.Length(min=1, max=MAX_CONNECTIONS),
+            ),
             cv.Optional("device_info", default=[]): cv.ensure_list(
                 cv.Schema(
                     {
@@ -48,22 +63,23 @@ CONFIG_SCHEMA = (
         }
     )
     .extend(esp32_ble_tracker.ESP_BLE_DEVICE_SCHEMA)
-    .extend(cv.COMPONENT_SCHEMA)
+    .extend(cv.COMPONENT_SCHEMA),
+    validate_connections,
 )
 
 
 async def to_code(config):
     var = cg.new_Pvariable(config[CONF_ID])
     await cg.register_component(var, config)
+
     await esp32_ble_tracker.register_ble_device(var, config)
 
-    connection_var = cg.new_Pvariable(config["connection"][CONF_ID])
-    cg.add(connection_var.set_mesh_name(config["mesh_name"]))
-    cg.add(connection_var.set_mesh_password(config["mesh_password"]))
+    cg.add(var.set_mesh_name(config["mesh_name"]))
+    cg.add(var.set_mesh_password(config["mesh_password"]))
 
     for device in config.get("device_info", []):
         cg.add(
-            connection_var.register_device(
+            var.register_device(
                 device["device_type"],
                 device["product_id"],
                 device["name"],
@@ -73,7 +89,11 @@ async def to_code(config):
             )
         )
 
-    await cg.register_component(connection_var, config["connection"])
-    cg.add(var.register_connection(connection_var))
-    cg.add(var.set_address_prefix(config["address_prefix"]))
-    await esp32_ble_tracker.register_client(connection_var, config["connection"])
+    if config.get("address_prefix"):
+        cg.add(var.set_address_prefix(config["address_prefix"]))
+
+    for connection_conf in config.get(CONF_CONNECTIONS, []):
+        connection_var = cg.new_Pvariable(connection_conf[CONF_ID])
+        await cg.register_component(connection_var, connection_conf)
+        cg.add(var.register_connection(connection_var))
+        await esp32_ble_tracker.register_client(connection_var, connection_conf)
